+++ packages/react/src/patterns/OneDataCollection/visualizations/collection/EditableTable/types.ts
@@ -125,6 +125,9 @@ export type SelectCellConfig<R extends RecordType> = {
           GroupingDefinition<RecordType>
         >,
         | "selectable"
+        | "selectionDisabled"
+        | "selectionInherited"
+        | "disableSelectAll"
         | "grouping"
         | "defaultGrouping"
         | "currentGrouping"