+++ packages/react/src/ui/checkbox.tsx
@@ -33,8 +33,8 @@ const Checkbox = React.forwardRef<
           aria-label={props.title}
           className={cn(
             "relative h-6 w-6 shrink-0 rounded-sm text-f1-foreground-selected data-[state=checked]:text-f1-foreground-inverse",
-            "after:absolute after:left-0.5 after:top-0.5 after:z-[1] after:h-5 after:w-5 after:rounded-xs after:border after:border-solid after:border-f1-border after:transition-[background-color,border-color] after:content-[''] hover:after:border-f1-border-hover data-[state=checked]:after:bg-f1-background-selected-bold hover:data-[state=checked]:after:border-transparent",
-            disabled && "cursor-not-allowed opacity-50 hover:border-f1-border",
+            "after:absolute after:left-0.5 after:top-0.5 after:z-[1] after:h-5 after:w-5 after:rounded-xs after:border after:border-solid after:border-f1-border after:transition-[background-color] after:content-[''] data-[state=checked]:after:bg-f1-background-selected-bold",
+            disabled && "cursor-not-allowed opacity-50",
             indeterminate && "data-[state=checked]:text-f1-foreground-inverse",
             props.checked &&
               disabled &&